// teams/18-Avault/src/avault-shiden-web/src/config/constants/types.ts
export interface Address {
  56?: string;
  81?: string;
  82?: string;
  87?: string;
  336?: string;
  592?: string;
}
export interface Token {
  symbol?: string;
  address?: Address;
  decimals?: number;
  projectLink?: string;
  busdPrice?: string;
}
// ...
export interface FarmConfig {
  pid: number;
  lpSymbol: string;
  lpAddresses: Address;
  lpMasterChefes: Address;
  token: Token;
  quoteToken: Token;
  decimals: number;
  lpDetail: {
    symbol: string;
    address: Address;
    decimals: number;
  };
  multiplier?: string;
  isCommunity?: boolean;
  dual?: {
    rewardPerBlock: number;
    earnLabel: string;
    endBlock: number;
  };
}